+++ app/controllers/fedipub/server_controller.rb
@@ -6,6 +6,7 @@ module Fedipub
     include Pundit::Authorization
     include Fedipub::ServerHelper
 
+    before_action :verify_request_signature!
     after_action :verify_authorized
 
     protect_from_forgery with: :null_session
@@ -16,7 +17,17 @@ module Fedipub
                 Fedipub::DataEntity::TombstonedError,
                 with: :error_gone
 
+    def self.require_signature?
+      false
+    end
+
     private
+
+    def verify_request_signature!
+      Fediverse::Signature.verify!(request: request, require_signature: ServerController.require_signature?)
+    rescue Fediverse::Signature::BadSignature
+      head :unauthorized
+    end
 
     def error_fallback(exception, fallback_message, status)
       message = exception&.message || fallback_message