+++ spec/requests/federation/followings_spec.rb
@@ -23,6 +23,17 @@ RSpec.describe '/federation/followings', type: :request do
       expect(response).to be_successful
     end
 
+    it 'rejects badly-signed requests' do
+      get fedipub.server_actor_following_url(actor, following), headers: { accept: Mime[:activitypub], signature: 'poop' }
+      expect(response).to have_http_status :unauthorized
+    end
+
+    it 'rejects unsigned requests when signatures are required' do
+      allow(Fedipub::ServerController).to receive(:require_signature?).and_return(true)
+      get fedipub.server_actor_following_url(actor, following), headers: { accept: Mime[:activitypub] }
+      expect(response).to have_http_status :unauthorized
+    end
+
     ACTIVITYPUB_CONTENT_TYPES.each do |accept|
       it "responds with LD in response to a #{accept} request" do
         get fedipub.server_actor_following_url(actor, following), headers: { accept: accept }